Fix signed/unsigned comparison in inferred species loop
Agent-Logs-Url: https://github.com/zhyan0603/GPUMD/sessions/88601893-e60a-4fce-a8f1-3a17af754198

+++ src/model/read_xyz.cu
@@ -352,9 +352,9 @@ void read_xyz_in_line_3(
 
     if (infer_type_from_xyz) {
       int type_id = -1;
-      for (int t = 0; t < atom_symbols.size(); ++t) {
+      for (size_t t = 0; t < atom_symbols.size(); ++t) {
         if (cpu_atom_symbol[n] == atom_symbols[t]) {
-          type_id = t;
+          type_id = (int)t;
           break;
         }
       }
